Changed coordinate to double.

+++ src/data/Coordinate.java
@@ -4,10 +4,10 @@ package data;
  * Created by Wesley on 6/02/2015.
  */
 public class Coordinate {
-	public int x;
-	public int y;
+	public double x;
+	public double y;
 
-	public Coordinate(int x, int y) {
+	public Coordinate(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
