+++ src/vision/PixelGroup.java
@@ -21,7 +21,7 @@ public class PixelGroup {
 		mostBottomCorner = new Coordinate(x, y);
 	}
 
-	public int getSize() {
+	public double getSize() {
 		return (mostBottomCorner.y - mostTopCorner.y) + (mostRightCorner.x - mostLeftCorner.x);
 	}
 
@@ -40,7 +40,7 @@ public class PixelGroup {
 		}
 	}
 
-	protected int squared (int x) {
+	protected double squared (double x) {
 		return x * x;
 	}
